requirements.txt
pymupdf==1.24.14
pytesseract==0.3.10
pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4

sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1

streamlit==1.37.1
python-dotenv==1.0.1

google-genai==0.6.0